LPS-92553 Rename classes to match implementation
