+++ modules/apps/asset/asset-display-impl/src/main/java/com/liferay/asset/display/internal/contributor/AssetEntrySummaryAssetInfoDisplayContributorField.java
@@ -14,14 +14,10 @@
 
 package com.liferay.asset.display.internal.contributor;
 
-import com.liferay.asset.display.contributor.AssetDisplayContributorField;
+import com.liferay.asset.display.contributor.AssetInfoDisplayContributorField;
 import com.liferay.asset.kernel.model.AssetEntry;
-import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.util.FastDateFormatFactoryUtil;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-
-import java.text.Format;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
@@ -33,14 +29,14 @@ import org.osgi.service.component.annotations.Component;
  */
 @Component(
 	property = "model.class.name=com.liferay.asset.kernel.model.AssetEntry",
-	service = AssetDisplayContributorField.class
+	service = AssetInfoDisplayContributorField.class
 )
-public class AssetEntryPublishDateAssetDisplayContributorField
-	implements AssetDisplayContributorField<AssetEntry> {
+public class AssetEntrySummaryAssetInfoDisplayContributorField
+	implements AssetInfoDisplayContributorField<AssetEntry> {
 
 	@Override
 	public String getKey() {
-		return "publishDate";
+		return "summary";
 	}
 
 	@Override
@@ -48,7 +44,7 @@ public class AssetEntryPublishDateAssetDisplayContributorField
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			locale, getClass());
 
-		return LanguageUtil.get(resourceBundle, "publish-date");
+		return LanguageUtil.get(resourceBundle, "summary");
 	}
 
 	@Override
@@ -58,14 +54,7 @@ public class AssetEntryPublishDateAssetDisplayContributorField
 
 	@Override
 	public String getValue(AssetEntry assetEntry, Locale locale) {
-		if (assetEntry.getPublishDate() == null) {
-			return StringPool.BLANK;
-		}
-
-		Format dateFormatDateTime = FastDateFormatFactoryUtil.getDateTime(
-			locale);
-
-		return dateFormatDateTime.format(assetEntry.getPublishDate());
+		return assetEntry.getSummary(locale);
 	}
 
 }
