+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/asset/display/contributor/DLFileEntryDownloadURLAssetInfoDisplayContributorField.java
@@ -12,13 +12,14 @@
  * details.
  */
 
-package com.liferay.adaptive.media.blogs.web.internal.asset.display.contributor;
+package com.liferay.document.library.web.internal.asset.display.contributor;
 
-import com.liferay.adaptive.media.content.transformer.ContentTransformerHandler;
-import com.liferay.adaptive.media.content.transformer.constants.ContentTransformerContentTypes;
-import com.liferay.asset.display.contributor.AssetDisplayContributorField;
-import com.liferay.blogs.model.BlogsEntry;
+import com.liferay.asset.display.contributor.AssetInfoDisplayContributorField;
+import com.liferay.document.library.util.DLURLHelper;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.repository.model.FileEntry;
+import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
 import java.util.Locale;
 
@@ -29,23 +30,21 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alejandro Tardín
  */
 @Component(
-	property = {
-		"model.class.name=com.liferay.blogs.model.BlogsEntry",
-		"service.ranking:Integer=2"
-	},
-	service = AssetDisplayContributorField.class
+	property = "model.class.name=com.liferay.document.library.kernel.model.DLFileEntry",
+	service = AssetInfoDisplayContributorField.class
 )
-public class AMBlogsEntryContentAssetDisplayContributorField
-	implements AssetDisplayContributorField<BlogsEntry> {
+public class DLFileEntryDownloadURLAssetInfoDisplayContributorField
+	implements AssetInfoDisplayContributorField<FileEntry> {
 
 	@Override
 	public String getKey() {
-		return "content";
+		return "downloadURL";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "content");
+		return LanguageUtil.get(
+			ResourceBundleUtil.getBundle(locale, getClass()), "download-url");
 	}
 
 	@Override
@@ -54,12 +53,17 @@ public class AMBlogsEntryContentAssetDisplayContributorField
 	}
 
 	@Override
-	public String getValue(BlogsEntry blogsEntry, Locale locale) {
-		return _contentTransformerHandler.transform(
-			ContentTransformerContentTypes.HTML, blogsEntry.getContent());
+	public String getValue(FileEntry fileEntry, Locale locale) {
+		try {
+			return _dlurlHelper.getDownloadURL(
+				fileEntry, fileEntry.getFileVersion(), null, StringPool.BLANK);
+		}
+		catch (Exception pe) {
+			return null;
+		}
 	}
 
 	@Reference
-	private ContentTransformerHandler _contentTransformerHandler;
+	private DLURLHelper _dlurlHelper;
 
 }
