+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/asset/display/contributor/DLFileEntryVersionAssetInfoDisplayContributorField.java
@@ -14,9 +14,7 @@
 
 package com.liferay.document.library.web.internal.asset.display.contributor;
 
-import com.liferay.asset.display.contributor.AssetDisplayContributorField;
-import com.liferay.document.library.util.DLURLHelper;
-import com.liferay.petra.string.StringPool;
+import com.liferay.asset.display.contributor.AssetInfoDisplayContributorField;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
@@ -24,27 +22,26 @@ import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import java.util.Locale;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Alejandro Tardín
  */
 @Component(
 	property = "model.class.name=com.liferay.document.library.kernel.model.DLFileEntry",
-	service = AssetDisplayContributorField.class
+	service = AssetInfoDisplayContributorField.class
 )
-public class DLFileEntryDownloadURLAssetDisplayContributorField
-	implements AssetDisplayContributorField<FileEntry> {
+public class DLFileEntryVersionAssetInfoDisplayContributorField
+	implements AssetInfoDisplayContributorField<FileEntry> {
 
 	@Override
 	public String getKey() {
-		return "downloadURL";
+		return "version";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
 		return LanguageUtil.get(
-			ResourceBundleUtil.getBundle(locale, getClass()), "download-url");
+			ResourceBundleUtil.getBundle(locale, getClass()), "version");
 	}
 
 	@Override
@@ -54,16 +51,7 @@ public class DLFileEntryDownloadURLAssetDisplayContributorField
 
 	@Override
 	public String getValue(FileEntry fileEntry, Locale locale) {
-		try {
-			return _dlurlHelper.getDownloadURL(
-				fileEntry, fileEntry.getFileVersion(), null, StringPool.BLANK);
-		}
-		catch (Exception pe) {
-			return null;
-		}
+		return fileEntry.getVersion();
 	}
-
-	@Reference
-	private DLURLHelper _dlurlHelper;
 
 }
