+++ modules/apps/journal/journal-web/src/main/java/com/liferay/journal/web/internal/asset/display/contributor/JournalArticleLastEditorNameAssetInfoDisplayContributorField.java
@@ -12,39 +12,43 @@
  * details.
  */
 
-package com.liferay.asset.display.internal.contributor;
+package com.liferay.journal.web.internal.asset.display.contributor;
 
-import com.liferay.asset.display.contributor.AssetDisplayContributorField;
-import com.liferay.asset.kernel.model.AssetEntry;
+import com.liferay.asset.display.contributor.AssetInfoDisplayContributorField;
+import com.liferay.journal.model.JournalArticle;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Jürgen Kappler
  */
 @Component(
-	property = "model.class.name=com.liferay.asset.kernel.model.AssetEntry",
-	service = AssetDisplayContributorField.class
+	property = "model.class.name=com.liferay.journal.model.JournalArticle",
+	service = AssetInfoDisplayContributorField.class
 )
-public class AssetEntryTitleAssetDisplayContributorField
-	implements AssetDisplayContributorField<AssetEntry> {
+public class JournalArticleLastEditorNameAssetInfoDisplayContributorField
+	implements AssetInfoDisplayContributorField<JournalArticle> {
 
 	@Override
 	public String getKey() {
-		return "title";
+		return "lastEditorName";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
-			locale, getClass());
+			locale, "com.liferay.journal.lang");
 
-		return LanguageUtil.get(resourceBundle, "title");
+		return LanguageUtil.get(resourceBundle, "last-editor-name");
 	}
 
 	@Override
@@ -53,8 +57,17 @@ public class AssetEntryTitleAssetDisplayContributorField
 	}
 
 	@Override
-	public String getValue(AssetEntry assetEntry, Locale locale) {
-		return assetEntry.getTitle(locale);
+	public String getValue(JournalArticle article, Locale locale) {
+		User user = _userLocalService.fetchUser(article.getUserId());
+
+		if (user != null) {
+			return user.getFullName();
+		}
+
+		return StringPool.BLANK;
 	}
+
+	@Reference
+	private UserLocalService _userLocalService;
 
 }
